refactor pipeline routes into separate router
